Submit Error Messages, extracting redudent application code
