+++ react-app/ccf/src/pages/application-form/subquestions/GrantProposal.tsx
@@ -1,5 +1,5 @@
-import { useState, useRef } from "react";
 import "./SubForm.css";
+import FileUploadSection from "../Components/FileUploadSection";
 
 type GrantProposalProps = {
   type: "Research" | "NextGen";
@@ -8,64 +8,6 @@ type GrantProposalProps = {
 };
 
 function GrantProposal({ type, formData, setFormData }: GrantProposalProps): JSX.Element {
-  const [selectedFile, setSelectedFile] = useState<File | null>(formData.file);
-  const [uploadStatus, setUploadStatus] = useState<string>("");
-  const fileInputRef = useRef<HTMLInputElement>(formData.file);
-
-  const handleFileChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (event.target.files && event.target.files[0]) {
-      console.log("File selected:", event.target.files[0].name);
-      setSelectedFile(event.target.files[0]);
-      setFormData({ ...formData, file: event.target.files[0] });
-      setUploadStatus("");
-    }
-  };
-
-  const handleDelete = () => {
-    console.log("Delete button pressed");
-    setSelectedFile(null);
-    setFormData({ ...formData, file: null });
-    setUploadStatus("");
-    if (fileInputRef.current) {
-      fileInputRef.current.value = "";
-    }
-  };
-
-  const handleUploadClick = () => {
-    console.log("Upload button clicked - opening file selector");
-    fileInputRef.current?.click();
-  };
-
-  const renderUploadSection = () => (
-    <div className="file-upload">
-      <label>Upload File (PDF Format)</label>
-      <br />
-      <br />
-      <div className="upload-btn-container">
-        <button 
-          className="upload-btn1" 
-          onClick={handleUploadClick}
-        >
-          {selectedFile ? selectedFile.name : "Click to upload"}
-        </button>
-        <input 
-          ref={fileInputRef}
-          type="file" 
-          accept=".pdf" 
-          onChange={handleFileChange}
-          style={{ display: "none" }} 
-          id="file-upload" 
-        />
-        {selectedFile && (
-          <button className="del-icon-container" onClick={handleDelete}>
-            <div className="del-icon"></div>
-          </button>
-        )}
-      </div>
-      {uploadStatus && <p className="upload-status">{uploadStatus}</p>}
-    </div>
-  );
-
   const renderResearchContent = () => (
     <div className="form-container">
       <div className="proposal-text">
@@ -88,7 +30,7 @@ function GrantProposal({ type, formData, setFormData }: GrantProposalProps): JSX
         </p>
         <br />
         <br />
-        {renderUploadSection()}
+        <FileUploadSection formData={formData} setFormData={setFormData} />
       </div>
     </div>
   );
@@ -122,7 +64,7 @@ function GrantProposal({ type, formData, setFormData }: GrantProposalProps): JSX
         </p>
         <br />
         <br />
-        {renderUploadSection()}
+        <FileUploadSection formData={formData} setFormData={setFormData} />
       </div>
     </div>
   );
