+++ react-app/ccf/src/pages/application-form/subquestions/NRNarrative.tsx
@@ -1,36 +1,8 @@
-import { useState, useRef } from "react";
 import "./SubForm.css";
 import { InformationProps } from "../../../types/application-types";
+import FileUploadSection from "../Components/FileUploadSection";
 
 function NRNarrative({ formData, setFormData }: InformationProps): JSX.Element {
-  const [selectedFile, setSelectedFile] = useState<File | null>(formData.file);
-  const [uploadStatus, setUploadStatus] = useState<string>("");
-  const fileInputRef = useRef<HTMLInputElement>(formData.file);
-
-  const handleFileChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (event.target.files && event.target.files[0]) {
-      console.log("File selected:", event.target.files[0].name);
-      setSelectedFile(event.target.files[0]);
-      setFormData({ ...formData, file: event.target.files[0] });
-      setUploadStatus("");
-    }
-  };
-
-  const handleDelete = () => {
-    console.log("Delete button pressed");
-    setSelectedFile(null);
-    setFormData({ ...formData, file: null });
-    setUploadStatus("");
-    if (fileInputRef.current) {
-      fileInputRef.current.value = "";
-    }
-  };
-
-  const handleUploadClick = () => {
-    console.log("Upload button clicked - opening file selector");
-    fileInputRef.current?.click();
-  };
-
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const { name, value } = e.target;
     setFormData((prevData: any) => ({ ...prevData, [name]: value }));
@@ -97,33 +69,7 @@ function NRNarrative({ formData, setFormData }: InformationProps): JSX.Element {
           className="text-input"
         />
         <br /><br /><br />
-        <div className="file-upload">
-          <label>Upload File (PDF Format)</label>
-          <br />
-          <br />
-          <div className="upload-btn-container">
-            <button 
-              className="upload-btn1" 
-              onClick={handleUploadClick}
-            >
-              {selectedFile ? selectedFile.name : "Click to upload"}
-            </button>
-            <input 
-              ref={fileInputRef}
-              type="file" 
-              accept=".pdf" 
-              onChange={handleFileChange}
-              style={{ display: "none" }} 
-              id="file-upload" 
-            />
-            {selectedFile && (
-              <button className="del-icon-container" onClick={handleDelete}>
-                <div className="del-icon"></div>
-              </button>
-            )}
-          </div>
-          {uploadStatus && <p className="upload-status">{uploadStatus}</p>}
-        </div>
+        <FileUploadSection formData={formData} setFormData={setFormData} />
       </div>
     </div>
   );
